+++ regression/run/cdesc/cdesc.cpp
@@ -10,3 +10,26 @@
 void Rank2In(int *arg)
 {
 }
+
+// Specialize template to emulate some sort of database which
+// returns different values based on type.
+template<>
+int getData<int>()
+{
+  return 1;
+}
+template<>
+long getData<long>()
+{
+  return 2;
+}
+template<>
+float getData<float>()
+{
+  return 3.0;
+}
+template<>
+double getData<double>()
+{
+  return 4.0;
+}